support large files by switching chunk offset to uint64

+++ cmd/plakar/cmd_info.go
@@ -29,7 +29,7 @@ import (
 )
 
 type JSONChunk struct {
-	Start  uint
+	Start  uint64
 	Length uint
 }
