+++ objects/objects.go
@@ -8,6 +8,6 @@ type Object struct {
 
 type Chunk struct {
 	Checksum [32]byte
-	Start    uint
+	Start    uint64
 	Length   uint
 }
